requirements.txt
# Slack
slack-bolt>=1.18.0
slack-sdk>=3.21.0

# LLM - Vertex AI with Claude
anthropic[vertex]>=0.25.0
google-cloud-aiplatform>=1.45.0

# Vector Store
chromadb<0.6.0

# Embeddings
sentence-transformers>=2.2.0

# Database
sqlalchemy>=2.0.0

# Web crawling
httpx>=0.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# GitHub API
PyGithub>=2.1.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0

# Numpy compatibility
numpy<2.0.0


